refactor(job): update job retrieval methods to use JobService and clean up imports
