+++ web-app/src/app/(app)/agents/[agentId]/jobs/@right/page.tsx
@@ -3,7 +3,7 @@ import { getTranslations } from "next-intl/server";
 
 import { CreateJobModalTrigger } from "@/components/create-job-modal";
 import { getSessionOrRedirect } from "@/lib/auth/utils";
-import { AgentService, getMyJobsByAgentId } from "@/lib/services";
+import { AgentService, JobService } from "@/lib/services";
 
 import JobDetailRedirect from "./components/job-detail-redirect";
 
@@ -22,14 +22,14 @@ export default async function RightSectionPage({
 
   const { agentId } = await params;
 
-  const agent = await AgentService.getInstance().getAgentById(agentId);
+  const agentService = AgentService.getInstance();
+  const agent = await agentService.getAgentById(agentId);
   if (!agent) {
     notFound();
   }
 
-  const agentJobs = await getMyJobsByAgentId(agentId);
-  const availableAgent =
-    await AgentService.getInstance().getAvailableAgentById(agentId);
+  const agentJobs = await JobService.getInstance().getMyJobsByAgentId(agentId);
+  const availableAgent = await agentService.getAvailableAgentById(agentId);
 
   if (agentJobs.length > 0) {
     return <JobDetailRedirect agentId={agentId} jobId={agentJobs[0].id} />;
