--- web-app/src/lib/services/job/index.ts
@@ -1,1 +1,0 @@
-export * from "./service";
